requirements.txt
torch
torchvision
timm
tqdm
msgpack
Pillow
datadings
numpy
torchprofile
fvcore
einops
PyWavelets
sinkhorn_transformer
routing_transformer
halonet_pytorch
